chore: improve FormControl stories
