+++ packages/chakra-ui-core/src/ColorModeProvider/index.js
@@ -1,2 +1,3 @@
 import ColorModeProvider from './ColorModeProvider'
 export default ColorModeProvider
+export * from './ColorModeProvider'
