+++ packages/chakra-ui-core/src/FormControl/FormControl.stories.js
@@ -15,13 +15,13 @@ storiesOf('UI | FormControl', module)
     components: { Input, FormControl, Stack, Icon, InputGroup, InputLeftElement, InputRightElement, FormHelperText },
     template: `
       <div>
-        <FormControl isInvalid id="test-error">
+        <FormControl id="test-error">
           <InputGroup>
             <InputLeftElement color="gray.300" fontSize="1.2em"><Icon name="lock" /></InputLeftElement>
             <Input :type="shouldShowPassword ? 'text' : 'password'" placeholder="Password" />
             <InputRightElement @click.native="shouldShowPassword = !shouldShowPassword" ><Icon :name="shouldShowPassword ? 'eye-slash' : 'eye'" color="gray.500" /></InputRightElement>
           </InputGroup>
-          <FormHelperText>Add your website here</FormHelperText>
+          <FormHelperText>Enter your password</FormHelperText>
         </FormControl>
       </div>
     `,
@@ -35,19 +35,20 @@ storiesOf('UI | FormControl', module)
     components: { Input, FormControl, Stack, Icon, InputGroup, InputLeftElement, InputRightElement, FormErrorMessage },
     template: `
       <div>
-        <FormControl isInvalid id="test-error">
+        <FormControl :isInvalid="value.length < 4" id="test-error">
           <InputGroup>
             <InputLeftElement color="gray.300" fontSize="1.2em"><Icon name="lock" /></InputLeftElement>
-            <Input :type="shouldShowPassword ? 'text' : 'password'" placeholder="Password" />
+            <Input v-model="value" :type="shouldShowPassword ? 'text' : 'password'" placeholder="Password" />
             <InputRightElement @click.native="shouldShowPassword = !shouldShowPassword" ><Icon :name="shouldShowPassword ? 'eye-slash' : 'eye'" color="gray.500" /></InputRightElement>
           </InputGroup>
-          <FormErrorMessage id="url-error">Website is invalid</FormErrorMessage>
+          <FormErrorMessage v-show="value.length < 4" id="url-error">Your password is too short.</FormErrorMessage>
         </FormControl>
       </div>
     `,
     data () {
       return {
-        shouldShowPassword: false
+        shouldShowPassword: false,
+        value: '123'
       }
     }
   }))
